Handle overlapping numbers
When the numbers are spelled out there's a possibility of overlap. Since
Rust's regex doesn't handle overlaps nicely the second possible digit is
ignored.

Rust doesn't support look-arounds another option is to reverse the
string and the regex, and then do a match.

This may actually be faster because we're also only looking for exactly
two matches, rather than all possible matches in the string.